requirements.txt
Flask==3.1.1
Werkzeug==3.1.3
openai>=1.99.9
python-dotenv==1.0.0
urllib3>=1.26.0
requests>=2.25.0
gunicorn==21.2.0
google-cloud-storage>=2.13.0
google-cloud-firestore>=2.12.0,<3
scikit-learn>=1.3.0
numpy<2
waitress==2.1.2
redis>=4.7.0
rq>=1.1.0